fix: remove postinstall hook
