+++ lib/utils.mjs
@@ -1,5 +1,7 @@
 /* eslint-disable max-classes-per-file, promise/param-names */
-const { red, cyan, green } = require('chalk');
+import pkg from 'chalk';
+
+const { red, cyan, green } = pkg;
 
 class Timeout {
   set(ms = 60 * 1000 * 5) {
@@ -15,7 +17,7 @@ class Timeout {
   }
 }
 
-async function withTimeout(action) {
+export async function withTimeout(action) {
   const timeout = new Timeout();
 
   return Promise.race([
@@ -26,7 +28,7 @@ async function withTimeout(action) {
   ]);
 }
 
-class Logger {
+export class Logger {
   constructor() {
     this.name = cyan('[bring-it]');
     this.OKAY = green('[okay]');
@@ -49,8 +51,3 @@ class Logger {
     console.log(this.name, key, ...message);
   }
 }
-
-module.exports = {
-  Logger,
-  withTimeout,
-};
